ARM: dts: msm: Add GMU CX GenPD instance for canoe GPU
Currently, there is a race condition in GenPD framework where
GPU CX GDSC can remain ON if both GMU and KGSL SMMU devices are
suspending in parallel and are voting on the same power domain.
Use a dedicated power domain for CX GDSC voting in canoe GPU
as per latest recommendation.

Change-Id: I62069f3809911ef3a4d6c9e1359582af532c9be0
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>