ARM: dts: msm: Add GMU CX GenPD instance for x1e80100 GPU
Currently, there is a race condition in GenPD framework where
GPU CX GDSC can remain ON if both GMU and KGSL SMMU devices are
suspending in parallel and are voting on the same power domain.
Use a dedicated power domain for CX GDSC voting in x1e80100 GPU
as per latest recommendation.

Change-Id: I4e30df1cd314e2d7bf541989b5f7ad644e7a178a
Signed-off-by: Pranjal Arya <pranarya@qti.qualcomm.com>