ARM: dts: msm: Update GMU frequency info and add synx memory node for Art
On the Art target, GMU scaling is on MX and not CX. As such, there
is no need for the qcom,gmu-perf-ddr-bw property in the DT. Update
the GMU frequency levels to reflect the latest guidance for the
Art target.

Additionally, add the synx-memory-region to allow synx memory to
be mapped to graphics firmware (as required by the synx feature).

Change-Id: I179101fd670c09b970011d4f3d2527f648dfbc4d
Signed-off-by: David Werner <daviwern@qti.qualcomm.com>